♻️ refactor: list/movie 컴포넌트 폴더 이동, 분리
